location_service/requirements.txt
greenlet==2.0.2
numpy==1.24.3
pandas==2.0.2
PyMySQL==1.0.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
six==1.16.0
SQLAlchemy==2.0.15
typing_extensions==4.6.3
tzdata==2023.3
